[482080] Move editors declarations in their own plugin
The commit reverts a part of bug 456771 (commit [1]).
For bug 456771, the editors declarations have been wrongly moved to
"ext" plugins. It did not make sense because, for example, having the
plugin oes.diagram.ui without the editor is useless. Better not include
the plugin oes.diagram.ui in his bundle if we do not want to.

The plugins o.e.s.table.ui.ext and o.e.s.tree.ui.ext are kept to allow a
correct update (even if they are empty now).

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=cd558cb06d7807291a78def287f13647e37676ab

Bug: 482080
Change-Id: I81653c1df5567d4551e00fcfbda6fe0e1d90f8e0
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>