[build] Remove problematic SWTBot feature from target platforms
The org.eclipse.swtbot.eclipse.test.junit we used to drag in our target
platforms depends on org.eclipse.core.runtime.compatibility.*, which are
being removed from Neon (they concern compatibility with Eclipse 2.x,
pre-OSGi). This means that our targets can not be resolved with Neon
M3 (and later). It looks like the SWTbot feature is not actually needed,
so remove it from all our targets (just for consistency).

Change-Id: I0d152f19e5aea183bec39ef66692feb0fe8268c5
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>