[findbugs] Fix broken SectionDescriptor initialization
The code was doing:

  afterSection = cfgElement.getAttribute(afterSection);

i.e. calling getAttribute() using an uninitialized/null field value, to
re-assign the result (null in this case) to the same field. So it was a
no-op.

From the context, what was probably meant was:

  afterSection = cfgElement.getAttribute("afterSection");

i.e. initialize the afterSection *field* with the value of the
correspondingly named *attribute* in the extension point configuration.

Change-Id: Ic2d708eac1ec5daea16705a9bf6595f99e0c2140
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>
Fast-tracked-by: Maxime Porhel <maxime.porhel@obeo.fr>