[482528] Support execution of ModelOperations outside of a DRepresentation
Add a variant of TaskHelper.buildTaskFromModelOperation() which does not
require a DRepresentation as part of the context information. This
allows the execution of ModelOperations in more contexts.

In particular, it allows SiriusToolsService to invoke callbacks
specified as ModelOperations for changes happening inside property
views, even if only given a semantic element (and not a
DSemanticDecorator).

Bug: 482528
Change-Id: I95d23b6760534c5b219337df2102ca99f388ceed
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>