[485136] Replace currentParticipant by currentParticipant()
currentParticipant is actually a service and should be denoted as such
in the AQL expressions. The lack of parenthesis makes AQL look for an
EStructural feature which does not exist and trigger many validation
errors.
At evaluation time the service is being used anyway.

Also replace "eContaienr"in some expressions by "eContainer()"

Bug: 485136
Change-Id: I39d7f3b0d520b71f28f6f37247a3eb19610bafdd