[482528] Do not wipe the FILES property when setting up SiriusInterpreter
This is a workaround relying on a specific concrete interpreter
implementation (ODesignGenericInterpreter). A better solution would be
to expose the getProperty() method directly on IInterpreter, but this
would break binary compatibility with the Query Legacy
interpreter (which is unmaintained), so a proper solution will have to
wait until #478878 is finished.

Bug: 482528
Change-Id: I74c36abb9642f0194414d7e6540327723a95c9ac
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>