[486828] Add UIState image
Set UIState.edit.image = true in viewpoint.genmodel. 
This property had an inconsistent value compared to the value of
DRepresentation.uiState.edit.createChild which set to true. 

This caused several NPEs on the selection of a DRepresentation from an
aird opened in the Sample Reflective Ecore Model Editor (for debugging
reasons)
 
Bug: 486828
Change-Id: I0c30694e77568f66362e7a79044e3ba3d334c731
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>