[478878] Merge the code and Test builds
Include the tests and their configuration into the base Maven build.
Remove the "test specific" targets platforms and update the "build"
target platform to include UML2.

The test specific build org.eclipse.sirius.tests.parent is not removed
but is made empty in order to be able to have Jenkins/hudson jobs which
can indiferently launch one layout or the other.

Bug: 478878
Change-Id: I9472ab5ee6a5d631cf054d5f809609d7567bec3d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>