[445371] Do not execute "DnD from ModelContent view" tests with Xvnc
There are 4 failing swtbot tests since the tests activation on Eclipse.
On Eclipse HIPP, the tests are launch through Xvnc. According to
analysis of Mickael Istria and Rohit Agrawal (Bug 285271), there is a
Xvnc limitation with DnD from SWTBot. So these tests were deactivated
when Xvnc is detected.
Tests are now executed with Junit4 to allow to ignore them when launched
with Xvnc. This "ignored" state is not visible when these tests are
executed through the suite. They are considered as "succeed" if
"ignored" because of the wrapping in Junit3 suite. If these tests are
launched individually, the "ignored" state is visible.

Bug: 445371
Change-Id: I4a336baff920747c13af63343152acbd553b40c1
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>