[test] Fix testWrapNodeInContainerLabelAtOpenning
On some CI server, the test
MultiLineLabelDiagramTest.testWrapNodeInContainerLabelAtOpenning()
fails. After analysis, we noticed that the same font is not exactly
interpreted in the same way according to OS.
"Comic Sans MS" has, most of the time, height = 17: ascent=13,
descent=4, leading=0.
But sometimes, when test failed, the height value is 15: ascent=12,
descent=3, leading=0.
This commit is not really a fix. It reduces the height of the figure to
have the same result in both condition.

Change-Id: I165d5b0ac60e320eb06eafbd423527298d847ea1
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>