[459993] TranslatableMessage Datatype
Addition of a TranslatableMessage datatype to the metamodel and
Viewpoint Specification String field that should be translatable have
been changed to this datatype.
Addition of a helper that will be in charge to select the proper
language (currently only returning the input value. This helper will be
completed in another commit.)

Bug: 459993
Change-Id: I0482a1ec5f8793d858ac8ab3b6706d898bfeef95
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>