[482528] Consider null as a valid evaluation result
It is perfectly legitimate for an expression to return null, but we
considered this as a failure which disabled a lot of behavior which
depends on IEvaluationResult.success().

Note that AQL raises a WARNING in its own EvaluationResult class when
a *service* returns null (instead of Nothing). It is debatable if this
is a good idea (see #487006), but for now this means we have to consider
diagnostics with a WARNING severity to represent successful evaluations.

Bug: 482528
Change-Id: Id7be86b5c330fd4e1d5dee5948f32a4aae32dd70
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>