[440296] Avoid to place a border node on an other side.
* If all parent sides are full, the border node is finally located on
the last conflicting position of the current side. But a conflict can be
detected with an other border node on an other side if its bounds
intersect the current border node we are trying to locate. If the
conflict is detected on an other side, we do not retain it for the final
location, except if there is no other choice.

Bug: 440296
Change-Id: I8787635e53475d449adc5f72a669081ceef060c5
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>