[489183] Add tests to reveal regressions of commit c8b5274d and e03deaf4
One of the goal of EdgeLayoutUpdaterModelChangeTrigger (and
RefreshEdgeLayoutNotificationFilter) is to used as much as possible the
figure to have precise computation.
But since commits [1] and [2], the behavior has changed and the figure
seems to be never used.
Two tests have been added to reveal these problems:
* testTgtChangingRoutingStyleWithGMFAutoSize: case of only one
notification
* testTgtChangingRoutingStyleFromStyleTabWithGMFAutoSize: case with two
notifications for the same change

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=c8b5274d3ea9459870ce652e15bcd8e359858d7d
[2]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=e03deaf4d64f55c7a56164ff075e118141c63861

Bug: 489183
Change-Id: Ifae2b8887e7d13276224b9aa9ea18e0a97c34e72
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>