[458751] Remove dependency from oes.ui to the JDT
The only actual code dependency left was in ModelingToggleNatureAction,
to be able to get the IProject from a JavaProject. However the way we
did this actually only depended on the fact that JavaProject implements
IProjectNature and has a getProject() method.

Using the more generic IProjectNature type makes the code more general
and allows us to remove the dependency to org.eclipse.jdt.core.

Bug: 458751
Change-Id: I4598ba18038605dc9b863a3077c90ecfc9e62b3e
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>