[489198] By-pass the update edge command in case of "layout"
If the source or the target of an edge has also been moved or resized,
we do not launch a specific command to re-layout edge. In this case we
consider that a full (or partial) layout has been done and that it is 
its responsibility to correctly set the edge layout (and respect the
centered edge notion).

Bug: 489198
Change-Id: I6a897c75459ddb5483f490e6aa5ef37a1e605b8f
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>