[490164] Give better feedback on failing session opening
- Have exception cause of failing session opening displayed in
ErrorDialog when session opening is done through
OpenRepresentationsFileJob.
- Now OpenRepresentationsFileJob manages only one session opening
instead
of a collection of session to open.
- Now to avoid concurrent execution of OpenRepresentationsFileJob we use
a specific ISchedulingRule defined in
AbstractRepresentationsFileJob.DEFAULT_SCHEDULING_RULE, instead of using
OpenRepresentationsFileJob.waitOtherJobs() which was provoking sometimes
reentrant call exception on IProject TreeItem.
- To have ProgressMonitorDialog displayed after an amount of time, we
call PlatformUI.getWorkbench().getProgressService().showInDialog() just
after OpenRepresentationsFileJob scheduling. In opposite to call of
PlatformUI.getWorkbench().getProgressService().busyCursorWhile() done
before in OpenRepresentationsFileJob.waitOtherJobs(), this allows to
have
ErrorDialog display the cause of failing session opening with
ProgressMonitorDialog.
- Now OpenSessionOnExpandListener hasn't need to call waitOtherJobs
since OpenRepresentationsFileJob cannot be executed in concurrency and
children of TreeItem appears then ExpandAgainJob is also useless.
- Remove Messages.OpenRepresentationsFileJob_initModelingProjectsTask
since OpenRepresentationsFileJob has no more need to compute main
representations files since it get representationsFileURI directly by
constructor, the case of ModelingProjects to manage is now done directly
in DefaultModelingProjectResourceListener.
- Catch always the RuntimeException only to create an error marker as
before.
- To avoid several OpenRepresentationsFileJob executed in sequence for a
same representationsFileURI, we call
ModelingProjectManager.loadAndOpenRepresentationsFile() which detects
that a job for a specified representationsFileURI is in execution,
instead of calling
OpenRepresentationsFileJob.scheduleNewWhenPossible().
- A new method has been added in ModelingProjectManager, a variant of
ModelingProjectManager.loadAndOpenRepresentationsFile(URI) with an
additional parameter to specify if the session opening job must be a
user job or not.
- To avoid "The resource tree is locked for modifications." message in
console and error log, due to markers deletion/addition not being done
in a workspace operation, we add InvalidModelingProjectMarkerUpdaterJob
WorkspaceModifyOperation.
- ModelingProject.getMainRepresentationsFileURI() no more update
workspace resource about markers as this must be done in a workspace
aware operation.

Bug: 490164
Change-Id: Iacbd8e06277bf70246976e84613d2d424edd9f28
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>