[461740] Correct navigation from a marker referencing external diagram
When a validation marker has been created from a DDiagram which does not
belong to the main aird, the marker references
. the main aird file (entry point of the session)
. the marked diagram uri (belongs to an other file)
. the marked element uri (used later for the selection)

This commit corrects the opening of such diagram belonging to a
referenced resource (which might not be a file) and ensure that when we
call DialectUIManager.INSTANCE.openEditor(session, dDiagram, pm), the
diagram instance we want to open is the one of the ResourceSet of the
current session. Previously, the passed dDiagram potentially came from
the ResourceSet of a temporary editor opened by Eclipse with a non-open
dummy session. This used to trigger the opening of the
DiagramDialectUIService.informOfActivateNeededViewpoints popup due to a
wrong analysis of activated Viewpoints the activateNeededViewpoint
method: the analyzed diagram coming from a phantom session.

Bug: 461740
Change-Id: I77b300c26bf8d1d96d19676f7551cc80e3e2cb73
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>