[459993] MessageTranslator updated for translation
- MessageTranslator now retrieves localized value of a String by
providing
the EObject or the bundle where the String is originating.
- MessageTranslator previously only needed the String value. Now the
EObject or the bundle where the String value comes from, is needed to
retrieve the localized value.
- Some calls to MessageTranslator have been removed because it should be
called from the UI. Direct access to the model element
TranslatableMessage fields still return the content of the field, not
the translation if available.

Bug: 459993
Change-Id: I516e1e8351f9febc2e68bcbc7113965947f52e50
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>