[442268] Relax Batik version ranges to allow for 1.7
Also add an explicit default SAX parser class name to work with Batik
1.7.

Batik seems to have a lot of trouble with even simple SVGs which can
cause TranscodingException. Assuming more recent versions have less of
these bugs/limitations, allow for 1.7 to be used when available instead
of limiting us to the older 1.6.

Bug: 442268
Change-Id: I3c48ee211432edbaf0f990b02ef49654d9613804
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>