[491604] Merge DRepresentationContainer into DView
DRepresentationContainer has been removed from the viewpoint.ecore
metamodel.

DRepresentationContainer used to extend DView with a single derived,
transient and volatile reference ("models"). In practice, it was the
only implementation of DView. 

DRepresentationContainer has been replaced by DView as serialized type. 
So a migration participant is needed for that, it will be provided in a
separated commit.

The "models" transient and derived reference has been moved to DView.
Given the nature of that reference it has no impact on the
serialization.
 
The features of DView have also been reordered to place the viewpoint
reference before the other multi valued containment references. This
will allow to serialize the reference to the viewpoint before the
contained representations.

Bug: 491604
Change-Id: I6821d1b76bb7fc8752ef75888fb77ca28db59533
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>