[492156] Fix memory leak at session closing
If one of the semantic resource is not in the ResourceSet of the
session, it is possible with a resource loaded from its URI http://xxxx,
the SemanticCrossReferencer is not removed from this resource at the
closing of the session.
With this commit, the DASI takes the same list of resources used during
SessionLazyCrossReferencer.initialize(), to remove the
SemanticCrossReferencer during the closing of the session.

Bug: 492156
Change-Id: I9e0b5af0209e55c71d5555761bd5ee2f73dd705e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>