[481846] Prevent newer Sirius model loading to avoid potential issues
 - We detect only the cases where a VSM or Representations resource has
version higher that the newest migration participant.
 - For Representations and VSM resources version mismatch, we close the
session when trying to open it. We provide a mechanism to overload this
behavior by providing you own implementation of
UICallBack.askSessionReopeningWithResourceVersionMismatch in which you
can have an UI feedback letting you choose if you want to open the
session anyway.
 - For VSM resource version mismatch, we prevent the viewpoint to be
register in the ViewpointRegistry (A message is added in the log)
 - When opening the VSM in its editor, natively, the odesign editor
displays the errors in the problems tab.
 - As the code getting and opening the session is mutualized for export
and aird open, the monitor work dividing up is defined once (30% for
getting session and 70% for opening session)
 - Tests are added to check VSM and aird version mismatch.
 
 - An exception thrown in the resource loading has been a considered but
was rejected. Consider the case when the resource is loaded because of
proxy resolution. In that cases, the exception is silently caught by
ECoreUtil.resolve. It is the case for referenced analysis and for the
viewpoint referenced by the View.
 
Bug: 481846 
Change-Id: I9e9209ba9b3a3d0ebdaf4661b969d56efa00cb45
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>