[494128] Encompass toggle nature action in WorkspaceModifyOperation
For a while, there are random failure on tests using the
ToggleNatureAction. Since the previous modification for bug 494128, the
ToggleNatureAction raises a dead lock. These 2 problems are linked. The
previous modification only reveals something wrong in
ToggleNatureAction.
So in this commit:
* from ModelingToggleNatureAction, the API calls to
removeModelingNature() and convertToModelingProject()
are now explicitly made in WorkspaceModifyOperation with Workspace
scheduling rule.
* in case of ModelingProjectManagerImpl.doAddModelingNature(), the
loadAndOpenRepresentationsFiles() is now called directly (a job is not
scheduled as this causes a deadlock).

Bug: 494128
Change-Id: I2324e5321b4bc33dad7d480a3079e2e8233d213f
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>