[493367] Make move, control uncontrol work
* move also the DRepresentationDescriptor when moving a DRepresentation
* In DRepQuery.getRepresentationDescriptor the way we get the session is
based on the DSemanticDecorator.target but in theory the DRepresentation
may not be a DSemanticDecorator. So this code is not robust. But we
can't base
on getExistingSession(representation.eResource().getURI()) neither
because, in case of controlled resource, that URI is not linked to the
session. We should improve the getSession(Resource) to make it work also
with Representation resource(not only semantic). That change is
impacting and should be the subject of another commit.


Bug: 493367
Change-Id: I1f4b47802b6ce2b8291bb47aa06e1d3a8ede2fa1
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>