[493367] Fix unwanted semantic resource loading at session opening
* At session opening, in
SessionResourcesTracker.initialize(IProgressMonitor), the resources that
have been loaded before forceLoadingOfEveryLinkedResource() are ignored
to be potentially added as semantic resource on the session. The problem
is that, because of
DRepresentationDescriptorCreationMigrationParticipant, some semantic
resources are loaded in the aird loading step because of DRepresentation
target proxy resolution. This commit avoids the proxy resolution to fix
the issue.

Bug: 493367
Change-Id: Icd84c3e0848b2fee0521609e93cb2ac61de0d202
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>