[497403] Compute GMF regions size according to container size
If the regions container has a predefined size, size defined in the VSM
(width and height computation expressions) or by the end-user at
creation, it is created with this predefined size. Its GMF layout
constraint is set with this predefined size.
During the first region creation, this size is reported to region and
the container size is set to {-1,-1}. Indeed, the size of the container
is computed from its children (except when it has no child).

Bug: 497403
Cherry-picked-from: 495119
Change-Id: Ie7a5f2cc8bee8448b9dd895a0c9e15bd67b4e73e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>