[499831] Expand the scope of the layout of regions container
This commit fixes the scenario of [1].
The previous commit 4c22f9 [2] reduces the number of regions container
layout. But a case was no longer handled: the deletion of a semantic
element represented by a region. In this case, the layout was not called
and the remaining regions are incorrectly layouted.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=495707#c16
[2]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=4c22f9b8234cafc16a93dea18dfe491194571342

Bug: 499831
Cherry-picked-from: 495707
Change-Id: Ideb7bddffa044adae5782d585dbe20c620c634da
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>