[495119] Fix pb of collapsing last region in horizontal stack
Since previous commit for bug 495119 (Compute GMF regions size according
to container size), there is a regression on collapsing the last region
in horizontal stack (a blank space remains if the regions container is
not in auto-size).

In AirXYLayoutEditPolicy [1], the constraint of a regions container is
force to "-1" for height (see commit 77d6b284 and method
getConstraintFor). This is for facility reason to handle the case when
the label is hidden/shown. This "-1" explains the difference between
HStack
and VStack when collapsing the last region. For the same facility
reason, the width is also constraint to "-1" here. A clean way would be
to set this "-1" directly during the resize command.

[1]
org.eclipse.sirius.diagram.ui.graphical.edit.policies.AirXYLayoutEditPolicy

Bug: 495119
Change-Id: I5b9e537bb14bd87c449bb1311456b0c22ba62225
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>