[499828] Change Font used for data test case of compartments tests.
This commit replaces the fonts used for data test case of compartments
tests by "Commic Sans MS". This font has similar size on different OS
(Windows, Linux, Mac). This allows to have reliable results on
DDiagramElements size when they are based on the label size.

Even with this font, there are some differences. For example, delta of
one pixel in height between a local Windows 10 and Eclipse Hipp on
Linux. So the test is adapted to not really checks height.

Bug: 499828
Cherry-picked-from: 495119
Change-Id: Ide681715a2f12d834281d2ff9585c737586260ac
Signed-off-by: Jessy Mallet <jessy.mallet@obeo.fr>