[496532] extend IRefreshExtension to allow for overriding the refresh
Introduce an interface IRefreshOverride which can be implemented by
IRefreshExtension instances to override the refresh default
implementation.

This mechanism could have been made more general but we don't envision
many usages besides very specific or experimental contexts.
Because of this we won't expose this new possibility in the release
notes.

Bug: 496532
Change-Id: Idc1f382ad1a60e778365f8c5953eb7655695d27b
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>