[499829] Fix wrong GMF size after region insertion
The result with this fix is not really correct, but it is better than
the current dimension computed from the location of the next region (for
wrong reason).

Also backport: Fix CCE and logic bug in GMFHelper.getGraphicalEditPart()
and also use more precise variable names, and return early when the
editor we're looking for is found.

Bug: 499829
Cherry-picked-from: 496744
Cherry-picked-from: 485426
Change-Id: Id33cba6263ed9438c907a42f249e19a6f703944b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>