[499106] Fix pb of png gradient according to figure location
The export of png with gradient was KO. According to the location of the
shape, the gradient was (or was not) correctly drawn.
The previous commit [1] does not consider the translation offset of the
SiriusGraphicsToGraphics2DAdaptor.
This commit also fixes some gradient computations (confusion between x,y
of second point and width,height).

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=225f5107a327a9817035ba761203fdb5f93bedd2

Bug: 499106
Change-Id: I991c92086978d1144814ab1a63c7f0a3afaea3ac
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>