[498341] Fix bad feedback when dragging bordered node in layouting mode
When layouting mode is enabled, the drop feedback is displayed but
nothing is done during the drop command execution.

Now we only have feedback around the container parent even if the mouse
is over another container. 

The drop command is now unexecutable, this re-enable the move and its
feedback on the parent. When the user relase the mouse button, the move
is done as shown by the feedback.

This change currently affects only border nodes. Additional work would
be needed in feedback and command computation to correctly handle nodes
contained in containers or diagram.

Change-Id: I6b7eeee77b5e7ae76448fe7396ab4e51bf29e439
Signed-off-by: pguilet <pierre.guilet@obeo.fr>