[498914] Consider the new bgPattern attribute in popState and pushState
In bug 491913, a new attribute bgPattern has been added to correctly
handled the gradient in all kind of format for export as image. But it
is not correctly set during the popState() and pushState() methods.
So after drawing a shape with gradient background color, there was a
side effect for following shapes.

Bug: 498914
Bug: 491913
Change-Id: I17d7b370f4581ad4abe5b3545b380f99e310f522
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>