[496466] Correctly handle custom features list of Sirius style copy
In the case of a Paste Style (or Paste Format), the custom features list
was not correctly handled. For example, if the source style had no
custom feature, when it was applied to the target shape, the copy of the
style had no custom feature even if some of them were different from the
original style of the target shape.
With this problem it is not possible to "Reset style properties to
default values".

Bug: 496466
Change-Id: I63a048547a44615271b403c0a27d438d30a59f8c
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>