[498341] Fix bad feedback when dragging node in layouting mode
When layouting mode is enabled, the drop feedback for simple node or for
container was still displayed but nothing was done during the drop
command execution.

Now, when a node is dragged outside of its container, the forbidden
figure will be displayed and it will not be shifted when dropping.

A new query has been added in EditPartQuery to check the layoutEditMode
state on an EditPart and used in existing code.

Bug: 498341
Change-Id: I8597889dd043049bbd9c8ba080e692818b2ff187
Signed-off-by: pguilet <pierre.guilet@obeo.fr>