[perf] Directly reuse the mapping of the candidate.
MappingsUpdater.updateMappings() is called before the creation of the
AbstractDNodeCandidate/DEdgeCandidate elements used during mapping
refreshes. So we can reuse the mapping of the candidate instead of
asking the mapping of the element once again. In some situations like
the DDiagramElement being a CDOObject and the mapping a local object,
this can become costly on big diagrams.

Change-Id: I22fd50902bbed5bc7ffe83c0fb48b4c794e0a145
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>