[499991] Clarify what is done with Layout menu
Before adding new elements in this menu for bug 499991, this commit
clarifies what is done (with comments) and makes some changes.
Explanation: The Layout menu that appears next to Format menu was
initially provided by GMF under the Format menu and called Arrange menu.
Change made in this commit: The pin/unpin actions are now directly
created in Arrange menu, this avoids an additional menu "refactoring"
later in DiagramEditorContextMenuProvider.

Bug: 499991
Change-Id: I3b6d90f001815846421f534d93fa141051c68564
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>