[496006] Make executeOperation service internal
Change the signature of the STS.executeOperation service so that it is
not invoked on the semantic element (which is a general EObject), but on
an instance of the private SiriusInputDescriptor type (typically the
"input" variable). This makes it impossible for the service to clash
with user-supplied services which would have compatible signatures.

Bug: 496006
Change-Id: Icc3cffe8c40c3ac89cc590e5f437b763cd2d8527
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>