[501515] Remove the Sirius ECrossRefAdapter from deleted representations
The ECrossReferenceAdapter is not removed from an object being deleted
if this object is a root object of the resource, which is now the case
for DRepresentations (since bug #494766).

This commit overloads the selfAdapt method in SessionLazyCrossReferencer
to remove the cross-referencer from DRepresentations (and only them)
when they are removed from the root of a Resource. The normal EMF
behavior is not changed for other cases.

Bug: 501515
Change-Id: Iaec65f86e17a6e67bff708ad4fb0ae0866b74361
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>