[502048] Improve diagnostics for AQL evaluation errors
We create a BasicDiagnostic to hold up to two children. In most cases it
will have only one, but unless client code explicitly digged into it it
would get an empty Diagnostic.getMessage() (from the wrapper) instead of
the details from the only actual child.

Instead, explicitly unwrap the contained diagnostic if there is only one
so that its full details are directly exposed.

Bug: 502048
Change-Id: I34ae17b5a6e21c666c3137667e37c9a048e26a3c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>