[496006] Add EditSupport type in the metamodel
The new type only defines EOperations with signatures matching the
SiriusToolServices#eefView* service methods. No concrete implementation
is provided (as it requires dependencies on oes.ui.properties and EMF
Edit to implement properly), so the raw EClass should not be used and is
only defined in the metamodel for the benefit of AQL, which will see the
EOperations and allow them to be invoked.

Bug: 496006
Change-Id: I8bbe4e7912054c2ca66534968baa06fe2497000d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>