[496006] Use EditSupport operations in default rules model
Migrate the default rules to use operations defined in EditSupport
instead of the publicly visible SiriusToolServices#eefView* services.

Note that for the eGet() service, there is no equivalent operation
needed in EditSupport, as AQL already provides this (although it needs a
small adjustement as AQL takes the feature name as argument, not the
EStructuralFeature itself).

Bug: 496006
Change-Id: I4a6d541908aa4af138f889caf8fe9d6dc9d9739d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>