[496058] Unwrap the EEFViewCategory when appropriate
EEFViewCategory is used to "tag" the actual semantic EObject with a
category name, but can cause issue in some cases as the "self" EObject
seen by some code is not what they expect. Unwrap it before calling
external helper methods so that they see the actual underlying
EObject (with the correct type).

Bug: 496058
Change-Id: I2c39a3464d8bd43eb9edceed33e6faa453a67fe3
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>