[496061] Rework the solution to integrate the default rules in the main sub-menu
The previous approach used the "menuBuilder" extension point to
contribute the action, but it had the unfortunate side-effect of showing
the new menu as a first entry. It becomes very confusing for users which
are accustomed to see the "New > Viewpoint" as the first entry in the
first context sub-menu.

This new approach relies on the EMF ChildExtender mechanism instead, to
contribute the default rules in addition to the "plain"
ViewExtensionDescription instance. Because of the way menu labels are
computed, by default both would have the same label, so we use a
temporary internal adapter to mark the default rules instance with a
custom label, which is used if available in
AbstractMenuBuilder.CustomCreateChildAction.

Bug: 496061
Change-Id: I3b08be24914dbb840f7585262ca486c0b0a09155
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>