[506322] Do not use ResourceStrategy to manage the Viewpoint registry
This reverts the part of de35505b28c6c6f5d583f3616950708e36b02b73 which
modified ViewpointRegistryImpl, and caused a regression when VSMs from
the workspace are deleted and a different version reappear at the same
location later.

The ResourceStrategy abstraction introduced for ticket #442133 was
designed for resources in the session's ResourceSet, but the
ViewpointRegistry maintains its own ResourceSet, with very specific
constraints on the resources' lifecycle to correctly handle
reload/refresh notifications. The abstraction is not even needed in this
case, as the performance problems it solves do not really occur here.

Bug: 506322
Change-Id: I03a798ee96e02096543a39b8223b25711d0c0fa2
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>