[506345] Add empty migration participant
This migration participant is used to artificially update the migration
version so that it is newer than the last migration participants of
property view. This is useful if the VSM has been created in "property
view" context and then is opened in a none "property view" context
because in that case, without this participant, the VSM would be
considered with a too recent version compared to the Sirius last
migration participant version.

Bug: 506345
Change-Id: If9cacfae3c06c3d9ff36be07e68a7fa2e48e0c51
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>