[506359] Authorize straighten actions for border node with several edges
The "straighten to" actions were previously disabled when an edge is
connected to border nodes that have several edges. It is now allowed.

The test has been updated with a case to illustrate the new authorized
case (but there is currently no specific checks for bendpoints stability
of other edges impacted by the straighten action). It will be commited
later in another commit.

Bug: 506359
Change-Id: Ib696a0d7b07f24a9418bc183b813139d1eb03302
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>