[501263] Fix memory leak in DDiagramEditorImpl
This applies to DDiagramEditorImpl the same approach as commit
bb18387550e27d3e0f5d41c94d35fc03730b517c did to the tree/table editors,
i.e. replace anonymous inner class (which has a hidden pointer to the
enclosing DDiagramEditorImpl) with a static internal class which cleans
it reference to the editor once it does not need it anymore.

Bug: 501263
Change-Id: I5d017a2708bdfbf146d754b3a93d99780991a302
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>