[build] Makes TargetPlatform ID more like an ID in the Oomph model
Oomph is now slightly more strict in the well formedness of 'id'
attributes in the model. This commit changes the value of the attribute
for the target platform task in a valid ID with no spaces in it. 

This could only impact external models which are referencing this
specific task but I don't see how that could be used by anybody out
there. On the other hand this non-conformity makes the whole Oomph
catalog not valid and Ed doesn't like having red crosses in his editor
;)


Change-Id: Ifeebf272b7b4fb12e988e038044c318f61c28e30