[513407] Ensure resource (re)loading happens in write transactions
* Resource loading was executed without transaction support; resource
reloading was executed in a read-only transaction. In case of languages
based on Xbase this caused problems, as Xbase modifies the models during
resolution (e.g. writes type information or similar steps). By moving
resource loading and reloading in write transactions, these issues are
handled correctly.
* Update the ValidateEditSupport used by Sirius
(ResourceModifiedFieldUpdater) to ensure that loaded resource are not
seen as modified.

Bug: 513407
Change-Id: I67c640fac8441647ae96005f265d1c2eb83548a6
Signed-off-by: Zoltan Ujhelyi <ujhelyiz@incquerylabs.com>
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>