[507056] Scale the GraphicContext to get better resolution at export
Compute a factor to reach the maximum size of image (specified through
the "_Pref_DiagramExportSizeMax" with a default value of 50 000 000
pixels)

This leads to a noticeable increase in image quality for most cases
while staying in limits preventing OutOfMemorry errors. This also has
the side effect of reducing the resolution when the diagram would have
been too big (which is probably better than just failing like right
now).

The method used to compute the factor applied is protected so that
implementers can override it, hence the code checking for an image which
is beyond the authorized limits is still there.


Bug: 507056
Change-Id: I2cf20f6c010e4b67b0c0c14d0628ed93f991e516