[506695] Rename drop source 'diagram' into 'representation' in the UI
The documentation and the VSM editor's UI used the "diagram" term to
denote the possible source of a drop tool, which actually handle all
types of Sirius representations.

Do not modify the metamodel type (which still has a DragSource.DIAGRAM
enum), but change how it is presented in the UI and update the
documentation to use the more general "representation" term.

Bug: 506695
Change-Id: Icc01d03e1c476bd594572149a2df83e74c122445
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>