[fix] Fix validation error in basicfamily sample VSM
Add a missing id to the tool section so that validating the VSM does not
produce an error. It still produces a warning because no metamodel is
associated, but we want this sample to be usable both deployed and from
the workspace, so registering the metamodel from the registry or the
workspace would produce errors/warnings in the other case anyway.

Change-Id: I5486f5dc30a2f791b397855eb5a460569431314d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>