[508435] Do not execute STS#executeOperation in a sub-transaction
Calls to STS#executeOperation are (normally) always performed in the
context of TEDCA.performModelChange(), which already opens an enclosing
transaction. Opening a sub-transaction here creates two operations on
the command stack for one top-level user operation, which messes with
undo/redo behavior.

Instead, always execute the SiriusCommand directly if it can be
executed. If it can't, we simply return without performing any actual
change. This still leaves an empty command on the stack in the case of
an unexecutable command, but properly cancelling the already started
enclosing transaction would require deeper changes that will have to
wait for a future version.

Bug: 508435
Change-Id: I0e7e966b67ebea1644f2d33cd69f3673eda9185f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>