[508969] Have "Font..." and "Font Color" contextual menus on edge name
With this commit, the "Font..." and "Font Color" contextual menus are
now available when an edge name is selected.

The plugin.xml has been modified to add "Font..." contextual menu for
AbstractDiagramNameEditPart.
The DiagramEditorContextMenuProvider.updateFormatMenu() has been
completed to add the menu "Font Color" when "Font..." is available. The
AbstractDEdgeNameEditPart has been modified to return the parent DEdge
for request corresponding to "Font Color" action.

Bug: 508969
Change-Id: Idf5276165be4a8b03f5e84501f0ade1dd7d75e28
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>